resume_ranking_api/requirements.txt
fastapi==0.110.0
uvicorn==0.27.0
python-multipart==0.0.9
openai==1.10.0
PyPDF2==3.0.1
python-docx==1.0.0
python-dotenv==1.0.0
pandas==2.1.1
pytest==7.4.3
httpx==0.26.0
